+++ astropy/table/table.py
@@ -18,7 +18,7 @@ from ..units import Quantity
 from ..utils import OrderedDict, isiterable, deprecated, minversion
 from ..utils.console import color_print
 from ..utils.metadata import MetaData
-from ..utils.data_info import InfoDescriptor, BaseColumnInfo, MixinInfo
+from ..utils.data_info import InfoDescriptor, BaseColumnInfo, MixinInfo, ParentDtypeInfo
 from . import groups
 from .pprint import TableFormatter
 from .column import (BaseColumn, Column, MaskedColumn, _auto_names, FalseArray,
@@ -45,9 +45,9 @@ def descr(col):
     This returns a 3-tuple (name, type, shape) that can always be
     used in a structured array dtype definition.
     """
-    col_dtype_str = col.dtype.str if hasattr(col, 'dtype') else 'O'
+    col_dtype = 'O' if (col.info.dtype is None) else col.info.dtype
     col_shape = col.shape[1:] if hasattr(col, 'shape') else ()
-    return (col.info.name, col_dtype_str, col_shape)
+    return (col.info.name, col_dtype, col_shape)
 
 
 def has_info_class(obj, cls):
@@ -436,6 +436,10 @@ class Table(object):
         def_names = _auto_names(n_cols)
 
         for col, name, def_name, dtype in zip(data, names, def_names, dtype):
+            # Structured ndarray gets viewed as a mixin
+            if isinstance(col, np.ndarray) and len(col.dtype) > 1:
+                col = col.view(NdarrayMixin)
+
             if isinstance(col, (Column, MaskedColumn)):
                 col = self.ColumnClass(name=(name or col.info.name or def_name),
                                        data=col, dtype=dtype,
@@ -875,6 +879,10 @@ class Table(object):
             # convert to a numpy array.
             if not hasattr(value, 'dtype') and not self._add_as_mixin_column(value):
                 value = np.asarray(value)
+
+            # Structured ndarray gets viewed as a mixin
+            if isinstance(value, np.ndarray) and len(value.dtype) > 1:
+                value = value.view(NdarrayMixin)
 
             # Make new column and assign the value.  If the table currently
             # has no rows (len=0) of the value is already a Column then
@@ -2211,3 +2219,45 @@ class QTable(Table):
             col = super(QTable, self)._convert_col_for_table(col)
 
         return col
+
+class NdarrayMixin(np.ndarray):
+    """
+    Minimal mixin using a simple subclass of numpy array
+    """
+    info = InfoDescriptor(ParentDtypeInfo)
+
+    def __new__(cls, obj, *args, **kwargs):
+        self = np.array(obj, *args, **kwargs).view(cls)
+        if 'info' in getattr(obj, '__dict__', ()):
+            self.info = obj.info
+        return self
+
+    def __array_finalize__(self, obj):
+        if obj is None:
+            return
+
+        if six.callable(super(NdarrayMixin, self).__array_finalize__):
+            super(NdarrayMixin, self).__array_finalize__(obj)
+
+        # Self was created from template (e.g. obj[slice] or (obj * 2))
+        # or viewcast e.g. obj.view(Column).  In either case we want to
+        # init Column attributes for self from obj if possible.
+        if 'info' in getattr(obj, '__dict__', ()):
+            self.info = obj.info
+
+    def __reduce__(self):
+        # patch to pickle Quantity objects (ndarray subclasses), see
+        # http://www.mail-archive.com/numpy-discussion@scipy.org/msg02446.html
+
+        object_state = list(super(NdarrayMixin, self).__reduce__())
+        object_state[2] = (object_state[2], self.__dict__)
+        return tuple(object_state)
+
+    def __setstate__(self, state):
+        # patch to unpickle NdarrayMixin objects (ndarray subclasses), see
+        # http://www.mail-archive.com/numpy-discussion@scipy.org/msg02446.html
+
+        nd_state, own_state = state
+        super(NdarrayMixin, self).__setstate__(nd_state)
+        self.__dict__.update(own_state)
+